Reorganise virtual robots and simulators
