+++ package.json
@@ -16,10 +16,10 @@
     "tslint:fix": "tslint -c tslint.json 'src/**/*.ts{,x}' --exclude 'src/shared/proto/**' --fix",
     "dev": "yarn start:sim",
     "start": "nodemon ./src/server/dev.ts",
-    "start:sim": "nodemon ./src/server/dev.ts --with-simulators",
-    "simulate": "nodemon ./src/simulators/simulate.ts",
+    "start:sim": "nodemon ./src/server/dev.ts --virtual-robots",
+    "virtual-robots": "nodemon ./src/virtual_robots/run.ts",
     "prod": "ts-node -F ./src/server/prod.ts",
-    "prod:sim": "ts-node -F ./src/server/prod.ts --with-simulators",
+    "prod:sim": "ts-node -F ./src/server/prod.ts --virtual-robots",
     "build": "yarn clean:build && webpack -p --progress --colors",
     "clean": "yarn clean:build && yarn clean:coverage",
     "clean:build": "rimraf build",
