+++ src/virtual_robots/flat_map.ts
