+++ src/virtual_robots/virtual_robot.ts
